Разбивка на модули, подключение вебпак, вебпак дев сервер
