+++ public/index.html
@@ -33,7 +33,7 @@
 
     <!-- Контент -->
   </main>
-  <script src="../src/main.js"></script>
+  <script src="./bundle.js"></script>
 </body>
 
 </html>
