+++ src/view/load-button.js
@@ -1,0 +1,3 @@
+const createLoadButtonTemplate = () => `<button class="load-more" type="button">load more</button>`;
+
+export {createLoadButtonTemplate};
